requirements.txt
streamlit
pandas
statsmodels
plotly
requests
joblib
aiohttp
holidays
skopt
Office365-REST-Python-Client
